+++ src/cli/javascript/mod.rs
@@ -3,12 +3,12 @@ pub use std::sync::Arc;
 
 use hyper::header::{HeaderName, HeaderValue};
 
-mod http_filter;
+mod request_filter;
 mod js_request;
 mod js_response;
 mod runtime;
 
-pub use http_filter::HttpFilter;
+pub use request_filter::HttpFilter;
 pub use js_request::JsRequest;
 pub use js_response::JsResponse;
 pub use runtime::Runtime;
